--- app/src/main/java/com/oneandonly/inventationblock/datasource/model/data/History.kt
@@ -1,7 +1,0 @@
-package com.oneandonly.inventationblock.datasource.model.data
-
-data class History(
-    var date: String,
-    var content: String,
-    var amount: String
-)
