--- app/src/main/java/com/oneandonly/inventationblock/datasource/model/data/Stock.kt
@@ -1,10 +1,0 @@
-package com.oneandonly.inventationblock.datasource.model.data
-
-data class Stock(
-    var name:String,
-    var stockCurrent: Int,
-    var stockSafe: Int,
-    var fixed: Boolean,
-    var unit: String,
-    var expired: Long
-)
