--- app/src/main/res/layout/item_main_list.xml
@@ -1,20 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<layout xmlns:android="http://schemas.android.com/apk/res/android">
-
-    <data>
-        <variable
-            name="ingredient"
-            type="com.oneandonly.inventationblock.datasource.model.data.Ingredient" />
-    </data>
-
-    <RelativeLayout
-        android:layout_width="match_parent"
-        android:layout_height="72dp"
-        android:layout_margin="4dp">
-
-
-
-
-
-    </RelativeLayout>
-</layout>
